backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
yfinance>=0.2.50
python-dotenv>=1.0.0
